requirements.txt
streamlit>=1.30.0
pyyaml>=6.0
openpyxl>=3.1.0
